Revert all changes which add the :meta protocol
