+++ lib/dalli/socket.rb
@@ -3,8 +3,6 @@
 module Dalli
   module Socket
     module InstanceMethods
-      READ_CHUNK_SIZE = 8196
-
       def readfull(count)
         value = String.new(capacity: count + 1)
         loop do
@@ -26,7 +24,7 @@ module Dalli
       def read_available
         value = +""
         loop do
-          result = read_nonblock(READ_CHUNK_SIZE, exception: false)
+          result = read_nonblock(8196, exception: false)
           if result == :wait_readable
             break
           elsif result == :wait_writable
@@ -38,58 +36,6 @@ module Dalli
           end
         end
         value
-      end
-
-      # Read a single \r\n-terminated line from the socket. Uses an internal
-      # buffer so that bytes read past the line boundary are preserved for
-      # subsequent read_line or read_from_buffer calls.
-      #
-      # @return [String] the next line including the trailing "\r\n"
-      def read_line
-        @read_buffer ||= +""
-        loop do
-          if (idx = @read_buffer.index("\r\n"))
-            return @read_buffer.slice!(0, idx + 2)
-          end
-          result = read_nonblock(READ_CHUNK_SIZE, exception: false)
-          case result
-          when :wait_readable
-            raise Timeout::Error, "IO timeout: #{safe_options.inspect}" unless IO.select([self], nil, nil, options[:socket_timeout])
-          when :wait_writable
-            raise Timeout::Error, "IO timeout: #{safe_options.inspect}" unless IO.select(nil, [self], nil, options[:socket_timeout])
-          when nil
-            raise Errno::ECONNRESET, "Connection reset: #{safe_options.inspect}"
-          else
-            @read_buffer << result
-          end
-        end
-      end
-
-      # Read exactly +count+ bytes from the socket, consuming from the
-      # internal read buffer first (populated by read_line overshoots).
-      #
-      # @param count [Integer] number of bytes to read
-      # @return [String] exactly +count+ bytes
-      def read_from_buffer(count)
-        @read_buffer ||= +""
-        while @read_buffer.bytesize < count
-          result = read_nonblock([count - @read_buffer.bytesize, READ_CHUNK_SIZE].max, exception: false)
-          case result
-          when :wait_readable
-            raise Timeout::Error, "IO timeout: #{safe_options.inspect}" unless IO.select([self], nil, nil, options[:socket_timeout])
-          when :wait_writable
-            raise Timeout::Error, "IO timeout: #{safe_options.inspect}" unless IO.select(nil, [self], nil, options[:socket_timeout])
-          when nil
-            raise Errno::ECONNRESET, "Connection reset: #{safe_options.inspect}"
-          else
-            @read_buffer << result
-          end
-        end
-        @read_buffer.slice!(0, count)
-      end
-
-      def clear_read_buffer
-        @read_buffer = nil
       end
 
       def safe_options
